DT: enhance cpu-map creation for SMT threads sharing resources
enabled the sharing of resources, such as L1 Cache and clocks, between SMT
threads. However, the cpu-map, pmu, coresight, etc. objects requires each
thread to point to a CPU node in order to be created properly. This patch
addresses that issue.

It introduces a fix that uses thread IDs to match each CPU thread in the
register array within the cpu-node. This ensures that the cpu-map is
correctly built even when SMT threads share resources.

Also, used a new variable in CPU node "**#cpu-cell**", in which I
describe the number of parameters when parsing the phandle with arg to
find the local thread ID. This vairable is not mandatary, and is indeed
backward-compatible.

Additionally, based on the discussions in [1], I adopted the new cpu-map
layout, where the first parameter is a phandle and the second is the
local thread index, as shown below:

    core0 {
      thread0 {
        cpu = <&cpu0 0>;
      };
      thread1 {
        cpu = <&cpu0 1>;
      };
    };

[1] https://lore.kernel.org/devicetree-spec/CAL_JsqK1yqRLD9B+G7UUp=D8K++mXHq0Rmv=1i6DL_jXyZwXAw@mail.gmail.com/

Signed-off-by: Alireza Sanaee <alireza.sanaee@huawei.com>